+++ modules/jaspr-code/src/extension.ts
@@ -7,7 +7,6 @@ import {
 
 import { createJasprProject, handleNewProjects } from "./create";
 import { jasprClean, jasprDoctor } from "./commands";
-import { ComponentCodeLensProvider } from "./code_lens";
 import { JasprServeDaemon } from "./jaspr/serve_daemon";
 import {
   findJasprProjectFolders,
@@ -16,6 +15,9 @@ import {
 import { JasprToolingDaemon } from "./jaspr/tooling_daemon";
 import { ScopesDomain } from "./jaspr/scopes_domain";
 import { HtmlDomain } from "./jaspr/html_domain";
+import { dartExtensionApi } from "./api";
+import { ComponentCodeLensProvider } from "./code_lenses/component_code_lens";
+import { ServeCodeLensProvider } from "./code_lenses/serve_code_lens";
 
 export async function activate(context: vscode.ExtensionContext) {
   let projects = await findJasprProjectFolders();
@@ -64,11 +66,13 @@ export async function activate(context: vscode.ExtensionContext) {
   );
 
   context.subscriptions.push(
-    vscode.commands.registerCommand("jaspr.serve", async () => {
-      const folderToRunCommandIn = await getFolderToRunCommandIn(
-        `Select the folder to run "jaspr serve" in`
-      );
-      if (!folderToRunCommandIn) {
+    vscode.commands.registerCommand("jaspr.serve", async (input: string | undefined, folder: string | undefined) => {
+      if (!folder) {
+        folder = await getFolderToRunCommandIn(
+          `Select the folder to run "jaspr serve" in`
+        );
+      }
+      if (!folder) {
         return;
       }
 
@@ -79,7 +83,8 @@ export async function activate(context: vscode.ExtensionContext) {
         name: "Jaspr",
         request: "launch",
         type: "jaspr",
-        cwd: folderToRunCommandIn,
+        cwd: folder,
+        args: !!input ? ['--input=' + input] : []
       });
     })
   );
@@ -97,6 +102,15 @@ export async function activate(context: vscode.ExtensionContext) {
       new ComponentCodeLensProvider(scopesDomain)
     )
   );
+
+  if (dartExtensionApi.version >= 3) {
+    context.subscriptions.push(
+      vscode.languages.registerCodeLensProvider(
+        { language: "dart", scheme: "file" },
+        new ServeCodeLensProvider()
+      )
+    );
+  }
 
   const htmlDomain = new HtmlDomain(toolingDaemon);
   context.subscriptions.push(htmlDomain);