+++ packages/jaspr_router/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.8.0
+
+- Using a `redirect` function on `Router` or `Route` while rendering on the server will now trigger a HTTP redirect (302) instead of rendering the target page.
+
 ## 0.7.3
 
 - `jaspr` upgraded to `0.21.1`